requirements.txt
captcha==0.4
cryptography==40.0.2
gTTS==2.3.2
langdetect==1.0.9
moviepy==1.0.3
opencv-python==4.7.0.72
Pillow==9.5.0
ping3==4.0.4
playsound==1.3.0
pytube==12.1.3
pywhatkit==4.6
requests==2.29.0
speedtest-cli==2.1.3